pasan las primeras 8

+++ finales/final2020-02-13/final-alumno/src/main/java/ar/uba/fi/objetovengers/Thanos.java
@@ -12,7 +12,7 @@ public class Thanos {
         ironMan.recibirAtaque(15);
     }
 
-    public void recibirAtaque(int puntosAtaque) {
+    public void recibirAtaque(double puntosAtaque) {
         this.puntosDeVida -= puntosAtaque;
     }
 }
